OCR model update
