+++ Ginger/GingerCore/GingerCore.csproj
@@ -198,6 +198,7 @@
     </Page>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Actions\ActOcr.cs" />
     <Compile Include="Actions\ActSikuli.cs" />
     <Compile Include="Actions\ActDemoPage.cs" />
     <Compile Include="Actions\ActImage.cs" />
@@ -246,7 +247,6 @@
     <Compile Include="Actions\UIAutomation\UIA_ActionRecorder.xaml.cs">
       <DependentUpon>UIA_ActionRecorder.xaml</DependentUpon>
     </Compile>
-    <Compile Include="ActOcr.cs" />
     <Compile Include="DataSource\DataSourceTableColumn.cs" />
     <Compile Include="Common\Devices\DeviceConfig.cs" />
     <Compile Include="Common\Devices\DeviceButton.cs" />
@@ -541,11 +541,10 @@
     <PackageReference Include="RestSharp">
       <Version>112.0.0</Version>
     </PackageReference>
-    <PackageReference Include="Sdcb.PaddleInference" Version="3.0.1" />
+    <PackageReference Include="Sdcb.PaddleInference" Version="3.0.0" />
     <PackageReference Include="Sdcb.PaddleInference.runtime.win64.mkl" Version="3.0.0.36" />
     <PackageReference Include="Sdcb.PaddleOCR" Version="3.0.0" />
-    <PackageReference Include="Sdcb.PaddleOCR.Models.Local" Version="3.0.1" />
-    <PackageReference Include="Sdcb.PaddleOCR.Models.Shared" Version="2.7.0.1" />
+    <PackageReference Include="Sdcb.PaddleOCR.Models.Online" Version="3.0.1" />
     <PackageReference Include="SharpSvn" Version="1.14001.156" />
     <PackageReference Include="SkiaSharp" Version="3.116.1" />
     <PackageReference Include="SkiaSharp.Views.Desktop.Common" Version="2.88.3" />
